+++ vllm_ascend/ops/triton/bincount.py
@@ -24,6 +24,7 @@ import torch
 from vllm.distributed.parallel_state import get_tp_group
 from vllm.triton_utils import tl, triton
 
+from vllm_ascend.ascend_config import get_ascend_config
 from vllm_ascend.ops.triton.triton_utils import get_vectorcore_num
 
 
@@ -80,10 +81,10 @@ def token_bin_counts_and_mask_kernel(
     )
 
     local_token = token - vocab_start_idx
-
     token_in_range = pos_mask & (token >= vocab_start_idx) & (local_token < vocab_size)
 
-    count_ptr = batch_counts_start + local_token * counts_vocab_stride
+    safe_local_token = tl.where(token_in_range, local_token, 0)
+    count_ptr = batch_counts_start + safe_local_token * counts_vocab_stride
     tl.atomic_add(count_ptr, 1, mask=token_in_range)
 
 
@@ -129,8 +130,11 @@ def get_token_bin_counts_and_mask_triton(
     progs = min(core_num, total_blocks)
     grid = (progs, triton.cdiv(total_blocks, progs))
 
-    tp_group = get_tp_group()
-    tp_rank = tp_group.rank_in_group
+    if get_ascend_config().enable_reduce_sample:
+        tp_group = get_tp_group()
+        tp_rank = tp_group.rank_in_group
+    else:
+        tp_rank = 0
     token_bin_counts_and_mask_kernel[grid](
         tokens,
         tokens.stride(0),
