+++ vllm_ascend/patch/worker/__init__.py
@@ -45,17 +45,7 @@ if not is_310p():
 else:
     import vllm_ascend.patch.worker.patch_idex_310  # noqa
 import vllm_ascend.patch.worker.patch_rejection_sampler  # noqa
-
-# TODO(MengqingCao): remove after the upstream community is modified
-import vllm_ascend.patch.worker.patch_llama_eagle3  # noqa
-
-# torchair/npugraph_ex is only available on NPU; silently skip when missing
-# so that CPU-only environments (e.g. UT runners without torch_npu) can still
-# import this module without crashing.
-try:  # noqa: SIM105
-    import vllm_ascend.patch.worker.patch_npugraph_ex_triton  # noqa
-except ImportError:
-    pass
+import vllm_ascend.patch.worker.patch_npugraph_ex_triton  # noqa
 import vllm_ascend.patch.worker.patch_kimi_k25  # noqa
 import vllm_ascend.patch.worker.patch_draft_quarot  # noqa
 import vllm_ascend.patch.worker.patch_cudagraph  # noqa
